Interval Arithmetic for plotting
The module implements basic binary operators for interval arithmetic.
The comparision of two intervals is handled through a three - valued
logic True, False and None.
The module is not perfectly accurate. It does not handle rounding up
and down of floating numbers. Hence it will face the same problems as
that of floating point arithmetic. The accuracy is sacrificed for speed.
Hence the module should be used only for plotting and not for interval
arithmetic operations. Use the mpmath interval arithmetic for interval
arithmetic calculations.