interval arithmetic handles domain and property tracking
The interval arithmetic has a variable is_valid which is three - valued
True - If the evaluation resulted in an valid interval. Also, the
evaluation resulted in an continuous function over the interval.
None - If part of the interval is not in the domain of the evaluating
function Or the function is not continuous in the interval.
False: The If the interval does not belong to the domain of the function
and hence cannot be evaluated.